requirements.txt
numpy==1.16.3
pandas==0.24.2
PyYaml==5.1
sklearn==0.0
h5py==2.9.0
keras==2.2.4
